+++ src/config/messages.config.js
@@ -10,10 +10,11 @@ const messagesData = {
 
       ]
     },
-    '小陶老师的爱徒666': {
+    '小陶的爱徒66': {
       teacherName: '小陶老师',
       teacherMessages: [
-
+        { id: 4, title: '22', content: '222', createdAt: '2026-04-22T02:12:33' },
+        { id: 3, title: '111', content: '111', createdAt: '2026-04-22T02:11:29' }
       ]
     },
     '跟成老师一起复习咯': {